+++ R/indifference.R
@@ -7,10 +7,11 @@
 #'
 #' @param .rebound_data A data frame of rebound information, 
 #'                      likely created by `rebound_analysis()`.
-#' @param graph_type See `ReboundTools::graph_types`.
-#' @param graph_params See `ReboundTools::graph_params`.
+#' @param graph_type See `ReboundTools::graph_types`. Default is `ReboundTools::graph_types$preferences`.
+#' @param graph_params Parameters that control the appearance of the graph. Default is `ReboundTools::default_graph_params`.
 #' @param q_dot_s_orig,C_dot_o_orig,f_Cs_orig,sigma See `ReboundTools::orig_vars`.
-#' @param q_dot_s_star,C_dot_o_star See `ReboundTools::star_vars`.
+#' @param q_dot_s_hat See `ReboundTools::hat_vars`.
+#' @param q_dot_s_bar,C_dot_o_bar See `ReboundTools::bar_vars`.
 #'
 #' @return A data frame of indifference curves for a preferences graph.
 #' 
@@ -29,8 +30,10 @@ indifference_lines <- function(.rebound_data,
                                f_Cs_orig = ReboundTools::orig_vars$f_Cs_orig,
                                sigma = ReboundTools::orig_vars$sigma,
                                
-                               q_dot_s_star = ReboundTools::star_vars$q_dot_s_star,
-                               C_dot_o_star = ReboundTools::star_vars$C_dot_o_star) {
+                               q_dot_s_hat = ReboundTools::hat_vars$q_dot_s_hat,
+
+                               q_dot_s_bar = ReboundTools::bar_vars$q_dot_s_bar, 
+                               C_dot_o_bar = ReboundTools::bar_vars$C_dot_o_bar) {
   
   meta <- extract_meta(.rebound_data)
   
@@ -40,40 +43,41 @@ indifference_lines <- function(.rebound_data,
   f_Cs0 <- .rebound_data[[f_Cs_orig]]
   sigma_val <- .rebound_data[[sigma]]
   
-  # Indifference curve at the orig point
+  # Indifference curve at the orig point (same as the star point, same as the hat point)
   qs1 <- qs0
   Co1 <- Co0
+  # Make sure we include the hat point in our curve.
+  qs2 <- .rebound_data[[q_dot_s_hat]]
+  
   qs1_qs0 <- qs1/qs0
   Co1_Co0 <- Co1/Co0
+  qs2_qs0 <- qs2/qs0
+  
   icurves <- add_indifference_curve(meta = meta,
                                     graph_type = graph_type,
                                     line_name = ReboundTools::rebound_stages$orig,
-                                    colour = graph_params$prefs_indiff_grid_colour,
-                                    size = graph_params$prefs_indiff_grid_size,
-                                    linetype = graph_params$prefs_indiff_grid_linetype,
                                     qs1_qs0 = qs1_qs0,
                                     Co1_Co0 = Co1_Co0,
+                                    qs2_qs0 = qs2_qs0,
                                     f_Cs_orig = f_Cs0,
-                                    sigma = sigma_val)
+                                    sigma = sigma_val, 
+                                    graph_params = graph_params)
   
-  # Indifference curve at the star point (after emplacement, before substitution)
-  # qs1 <- .rebound_data[[q_dot_s_star]]
-  # Co1 <- .rebound_data[[C_dot_o_star]]
-  # qs1_qs0 <- qs1/qs0
-  # Co1_Co0 <- Co1/Co0
-  # icurves <- icurves %>%
-  #   add_indifference_curve(meta = meta,
-  #                          graph_type = graph_type,
-  #                          line_name = ReboundTools::rebound_stages$orig,
-  #                          colour = grid_colour,
-  #                          size = grid_size,
-  #                          linetype = grid_linetype,
-  #                          qs1_qs0 = qs1_qs0,
-  #                          Co1_Co0 = Co1_Co0,
-  #                          f_Cs_orig = f_Cs0,
-  #                          sigma = sigma_val)
-  
-  
+  # Indifference curve at the bar point (after income effect)
+  qs1 <- .rebound_data[[q_dot_s_bar]]
+  Co1 <- .rebound_data[[C_dot_o_bar]]
+  qs1_qs0 <- qs1/qs0
+  Co1_Co0 <- Co1/Co0
+  icurves <- icurves %>%
+    add_indifference_curve(meta = meta,
+                           graph_type = graph_type,
+                           line_name = ReboundTools::rebound_stages$hat,
+                           qs1_qs0 = qs1_qs0,
+                           Co1_Co0 = Co1_Co0, 
+                           qs2_qs0 = NULL,
+                           f_Cs_orig = f_Cs0,
+                           sigma = sigma_val, 
+                           graph_params = graph_params)
   return(icurves)
 }
 
@@ -85,6 +89,14 @@ indifference_lines <- function(.rebound_data,
 #' The indifference curves are accumulated in rows.
 #' 
 #' The preferences graph is _always_ indexed, so there is no `indexed` argument.
+#' 
+#' This function finds a reasonable range over which to sweep the `qs_qs0` variable
+#' when generating the indifference curve, depending on the value of `qs2_qs0`.
+#' When `qs2_qs0` is `NULL` (the default), the curve is generated over the range
+#' `grqph_params$qs_qs0_lower * qs1_qs0` to `grqph_params$qs_qs0_upper * qs1_qs0`.
+#' When `qs2_qs0` is non-`NULL`, the curve is generated over the range
+#' `grqph_params$qs_qs0_lower * min(qs1_qs0, qs2_qs0)` to
+#' `grqph_params$qs_qs0_upper * max(qs1_qs0, qs2_qs0)`.
 #'
 #' @param .DF A data frame that accumulates preferences curves. 
 #'            When `NULL`, the default, a new data frame is created and returned.
@@ -94,14 +106,13 @@ indifference_lines <- function(.rebound_data,
 #' @param graph_type The graph type for the indifference curve.
 #'                   Default is `ReboundTools::graph_types$preferences`.
 #' @param line_name A name for this indifference curve.
-#' @param colour The colour for this indifference curve. 
-#'               Default is `ReboundTools::default_graph_params$prefs_indiff_curve_colour`.
-#' @param size Line width. Default is `ReboundTools::default_graph_params$prefs_indiff_grid_size`.
-#' @param linetype Line type. Default is `ReboundTools::default_graph_params$prefs_indiff_grid_linetype`.
 #' @param qs1_qs0,Co1_Co0 The (x,y) coordinates of a point on this indifference curve.
-#' @param f_Cs_orig The ration of spending on the energy service to 
+#' @param qs2_qs0 A second x value at which a a point on the indifference curve should be calculated. Default is `NULL`.
+#' @param f_Cs_orig The ratio of spending on the energy service to 
 #'                  the sum of initial spending on the energy service and other goods.
 #' @param sigma The elasticity of substitution between spending on the energy service and spending on other goods.
+#' @param graph_params Parameters that control the appearance of this graph.
+#' @param eeu_base_params See `ReboundTools::eeu_base_params`.
 #' @param graph_df_colnames A list of column names to use throughout the package.
 #'                          Default is `ReboundTools::graph_df_colnames`.
 #'
@@ -117,23 +128,58 @@ add_indifference_curve <- function(.DF = NULL,
                                    meta, 
                                    graph_type = ReboundTools::graph_types$preferences, 
                                    line_name, 
-                                   colour = ReboundTools::default_graph_params$prefs_indiff_grid_colour, 
-                                   size = ReboundTools::default_graph_params$prefs_indiff_grid_size,
-                                   linetype = ReboundTools::default_graph_params$prefs_indiff_grid_linetype,
-                                   qs1_qs0, Co1_Co0, f_Cs_orig, sigma,
+                                   qs1_qs0, Co1_Co0, qs2_qs0 = NULL, f_Cs_orig, sigma,
+                                   graph_params = ReboundTools::default_graph_params,
+                                   eeu_base_params = ReboundTools::eeu_base_params,
                                    graph_df_colnames = ReboundTools::graph_df_colnames) {
+  # Calculate x values at which indifference curve should be evaluated.
+  if (is.null(qs2_qs0)) {
+    min_qs <- qs1_qs0
+    max_qs <- qs1_qs0
+  } else {
+    min_qs <- lapply(X = list(qs1_qs0, qs2_qs0), FUN = min) %>% unlist()
+    max_qs <- lapply(X = list(qs1_qs0, qs2_qs0), FUN = max) %>% unlist()
+  }
+  x_vals <- Map(f = geom_seq, 
+                from = graph_params$qs_qs0_lower*min_qs, 
+                to = graph_params$qs_qs0_upper*max_qs, 
+                n = graph_params$n_indiff_curve_points)
+  # Be sure to include the qs1_qs0 and qs2_qs0 points, too.
+  x_vals <- Map(f = c, x_vals, qs1_qs0)
+  if (!is.null(qs2_qs0)) {
+    x_vals <- Map(f = c, x_vals, qs2_qs0)
+  }
+  x_vals <- lapply(X = x_vals, FUN = sort)
+  cases <- meta[[eeu_base_params$case]]
+  x_df <- Map(f = function(cas, x_v){data.frame(cas, x_v)}, cas = cases, x_v = x_vals) %>% 
+    dplyr::bind_rows() %>% 
+    magrittr::set_names(c(eeu_base_params$case, graph_df_colnames$x_col)) %>% 
+    # Eliminate unnecessary repeated rows
+    unique()
+  
   out <- meta %>% 
     dplyr::mutate(
       "{graph_df_colnames$graph_type_col}" := graph_type, 
       "{graph_df_colnames$line_name_col}" := line_name,
-      "{graph_df_colnames$colour_col}" := colour, 
-      "{graph_df_colnames$size_col}" := size,
-      "{graph_df_colnames$linetype_col}" := linetype,
-      qs1_qs0 = qs1_qs0,
-      Co1_Co0 = Co1_Co0,
-      f_Cs_orig = f_Cs_orig, 
-      sigma = sigma
-    )
+      "{graph_df_colnames$colour_col}" := graph_params$prefs_indiff_grid_colour, 
+      "{graph_df_colnames$size_col}" := graph_params$prefs_indiff_grid_size,
+      "{graph_df_colnames$linetype_col}" := graph_params$prefs_indiff_grid_linetype, 
+      "{graph_df_colnames$qs1_qs0_col}" := qs1_qs0,
+      "{graph_df_colnames$Co1_Co0_col}" := Co1_Co0,
+      "{graph_df_colnames$f_Cs_orig_col}" := f_Cs_orig, 
+      "{graph_df_colnames$sigma_col}" := sigma
+    ) %>% 
+    dplyr::full_join(x_df, by = eeu_base_params$case) %>% 
+    dplyr::mutate(
+      # Now add y values via indifference_func()
+      "{graph_df_colnames$y_col}" := indifference_func(qs_qs0 = .data[[graph_df_colnames$x_col]], 
+                                                       qs1_qs0 = .data[[graph_df_colnames$qs1_qs0_col]], 
+                                                       Co1_Co0 = .data[[graph_df_colnames$Co1_Co0_col]], 
+                                                       f_Cs_orig = .data[[graph_df_colnames$f_Cs_orig_col]], 
+                                                       sigma = .data[[graph_df_colnames$sigma_col]])
+    ) %>% 
+    # Get rid of any NaN values in the y column
+    dplyr::filter(!is.nan(.data[[graph_df_colnames$y_col]]))
   if (is.null(.DF)) {
     return(out)
   }
@@ -180,5 +226,34 @@ indifference_func <- function(qs_qs0, qs1_qs0, Co1_Co0, f_Cs_orig, sigma, rho = 
   term2 <- qs1_qs0^rho
   term3 <- qs_qs0^rho
   term4 <- Co1_Co0^rho
-  (term1*(term2 - term3) + term4)^(1/rho)
+  twominusthree <- term2 - term3
+  (term1*twominusthree + term4)^(1/rho)
 }
+
+
+#' Generate a geometric sequence
+#'
+#' @param from The first value in the sequence.
+#' @param to The final value in the sequence.
+#' @param n The number of points to include in the sequence.
+#' 
+#' @return A vector of numbers.
+#' 
+#' @export
+#'
+#' @examples
+#' geom_seq(from = 0.01, to = 10, n = 20)
+#' geom_seq(from = 10, to = 20, n = 30)
+geom_seq <- function(from, to, n) {
+  assertthat::assert_that(n >= 2, msg = "n >= 2 required in geom_seq()")
+  by = (to/from) ^ (1/(n-1))
+  from * by^(0:(n-1))
+}
+
+
+
+
+
+
+
+